requirements.txt
amqp==5.3.1
asgiref==3.8.1
asttokens==3.0.0
async-timeout==5.0.1
billiard==4.2.1
celery==5.5.3
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
Django==5.2.4
exceptiongroup==1.3.0
executing==2.2.0
idna==3.10
ipykernel==6.29.5
ipython==8.37.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kombu==5.5.4
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.2.6
packaging==25.0
pandas==2.3.0
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
psycopg==3.2.9
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==27.0.0
redis==6.2.0
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
tornado==6.5.1
traitlets==5.14.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
djangorestframework-simplejwt==5.3.1   # JWT token implementation for DRF
dj-rest-auth[with_social]==6.0.0       # REST API authentication utils